requirements.txt
python-dotenv=1.0.0
discord.py-self=2.0.0
spotipy=2.23.0